+++ CHANGELOG.md
@@ -4,4 +4,5 @@ All notable changes to _asdf-yamllint_ will be documented in this file.
 
 ## Changes
 
+- (2023-01-15) Fix `download` and `install` script errors on OSX.
 - (2023-01-14) Initial release.